+++ packages/runtime-vapor/src/apiDefineAsyncComponent.ts
@@ -5,7 +5,6 @@ import {
   createAsyncComponentContext,
   currentInstance,
   handleError,
-  isKeepAlive,
   markAsyncBoundary,
   performAsyncHydrate,
   useAsyncComponentState,
@@ -30,7 +29,6 @@ import { invokeArrayFns } from '@vue/shared'
 import { type TransitionOptions, insert, remove } from './block'
 import { parentNode } from './dom/node'
 import { setTransitionHooks } from './components/Transition'
-import type { KeepAliveInstance } from './components/KeepAlive'
 
 /*@ __NO_SIDE_EFFECTS__ */
 export function defineVaporAsyncComponent<T extends VaporComponent>(
@@ -151,10 +149,6 @@ export function defineVaporAsyncComponent<T extends VaporComponent>(
       load()
         .then(() => {
           loaded.value = true
-          // if parent is keep-alive, re-evaluate caching
-          if (instance.parent && isKeepAlive(instance.parent)) {
-            ;(instance.parent as KeepAliveInstance).ctx.onAsyncResolve(instance)
-          }
         })
         .catch(err => {
           onError(err)
@@ -174,7 +168,9 @@ export function defineVaporAsyncComponent<T extends VaporComponent>(
         }
 
         if (instance.$transition) frag!.$transition = instance.$transition
-        frag!.update(render)
+        frag.update(render)
+        // Manually trigger cacheBlock for KeepAlive
+        if (frag.keepAliveCtx) frag.keepAliveCtx.cacheBlock()
       })
 
       return frag